+++ src/configs/react.ts
@@ -1,10 +1,43 @@
+import { createRequire } from 'node:module';
+
 import type { TSESLint } from '@typescript-eslint/utils';
 import eslintReact from '@eslint-react/eslint-plugin';
 import stylisticPlugin from '@stylistic/eslint-plugin';
 import reactHooksPlugin from 'eslint-plugin-react-hooks';
 import jsxA11yPlugin from 'eslint-plugin-jsx-a11y';
+import type { PackageJson } from 'type-fest';
+import semver from 'semver';
 
 import { convertWarnsToErrorsIfNeeded } from './utils/convertWarnsToErrorsIfNeeded.ts';
+
+const require = createRequire(import.meta.url);
+const { version = '0.0.0' } = require('@eslint-react/eslint-plugin/package.json') as PackageJson;
+
+const getEslintReactRules = () => {
+  if (semver.satisfies(version, '^4.0.0')) {
+    return {
+      '@eslint-react/naming-convention-filename': 'error', // enforce corrent filename // v4 rename
+      // '@eslint-react/naming-convention/component-name': 'error', // TODO how to re-implement? was removed in v3 https://github.com/Rel1cx/eslint-react/releases/tag/v3.0.0-rc.0
+      // '@eslint-react/jsx-shorthand-boolean': 'error', // TODO how to re-implement? was removed in v4 https://github.com/Rel1cx/eslint-react/issues/1243
+      // '@eslint-react/jsx-shorthand-fragment': 'error', // TODO how to re-implement? was removed in v4 https://github.com/Rel1cx/eslint-react/issues/1243
+    } as const;
+  }
+  if (semver.satisfies(version, '^3.0.0')) {
+    return {
+      '@eslint-react/naming-convention/filename': 'error', // enforce corrent filename
+      // '@eslint-react/naming-convention/component-name': 'error', // TODO how to re-implement? was removed in v3 https://github.com/Rel1cx/eslint-react/releases/tag/v3.0.0-rc.0
+      '@eslint-react/jsx-shorthand-boolean': 'error',
+      '@eslint-react/jsx-shorthand-fragment': 'error',
+    } as const;
+  }
+  // ^2.0.0
+  return {
+    '@eslint-react/naming-convention/filename': 'error', // enforce corrent filename
+    '@eslint-react/naming-convention/component-name': 'error',
+    '@eslint-react/jsx-shorthand-boolean': 'error',
+    '@eslint-react/jsx-shorthand-fragment': 'error',
+  } as const;
+};
 
 const config: TSESLint.FlatConfig.ConfigArray = [
   {
@@ -63,19 +96,16 @@ const config: TSESLint.FlatConfig.ConfigArray = [
       'jsx-a11y/no-noninteractive-element-interactions': 'off', // TODO enable later
       'jsx-a11y/label-has-associated-control': 'off', // TODO enable later but with `assert`=`either`
 
-      '@eslint-react/naming-convention/filename': 'error', // enforce corrent filename
-
       // disable some recommended rules
       '@eslint-react/prefer-destructuring-assignment': 'off', // can break discriminated union types
 
       // enable airbnb-style rules
-      '@eslint-react/jsx-shorthand-boolean': 'error',
-      '@eslint-react/jsx-shorthand-fragment': 'error',
-      '@eslint-react/naming-convention/component-name': 'error',
       '@stylistic/jsx-curly-brace-presence': ['error', { props: 'never', children: 'never' }],
       '@stylistic/jsx-self-closing-comp': 'error',
 
       '@dartess/no-props-with-children-type': 'error',
+
+      ...getEslintReactRules(),
     },
   },
 ];